fusion_tools/optimization/sharpness_aware_evaluator.py
import numpy as np
from typing import List, Dict, Any

class SharpnessAwareEvaluator:
    """
    Applies Sharpness-Aware Minimization (SAM/ASAM) to evaluate model robustness and select models that generalize better.
    """
    def __init__(self):
        pass

    def evaluate(self, model_outputs: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        # Placeholder: In practice, would perturb inputs and measure output stability
        for output in model_outputs:
            # Simulate sharpness score (lower is better)
            output['sharpness_score'] = np.random.uniform(0, 1)
        # Sort by sharpness score
        return sorted(model_outputs, key=lambda x: x['sharpness_score']) 