fusion_tools/requirements.txt
requests>=2.31.0
flask>=2.3.0
pyyaml>=6.0
colorama>=0.4.6
rich>=13.5.0
fastapi>=0.104.0
uvicorn>=0.23.0
websockets>=11.0.0
aiohttp>=3.8.0
schedule>=1.2.0
psutil>=5.9.0 